% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{ab_fr}
\alias{ab_fr}
\title{ab_fr}
\usage{
ab_fr(dataset, col_name)
}
\arguments{
\item{dataset}{the dataset for which you wish to retrieve absolute frequency.}

\item{col_name}{the column name for which you wish to retrieve absolute frequency.}
}
\value{
returns the absolute frequency for a specified column.
}
\description{
ab_fr
}
\examples{
\dontrun{
ab_fr(dataset = soccer, col_name =  result_19_20)}
}
