% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{all_levels}
\alias{all_levels}
\title{all_levels}
\usage{
all_levels(dataset, col_name)
}
\arguments{
\item{dataset}{the dataset you wish to retrieve levels for}

\item{col_name}{the column name you wish to retrieve levels for}
}
\value{
returns levels of a categorical variable
}
\description{
all_levels
}
\examples{
\dontrun{
all_levels(dataset = soccer, col_name = result_19_20)}
}
