% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{cv_mode}
\alias{cv_mode}
\title{cv_mode}
\usage{
cv_mode(dataset, col_name)
}
\arguments{
\item{col_name}{the column name for which you wish to retrieve the mode}
}
\value{
returns levels of a categorical variable
}
\description{
cv_mode
}
\examples{
\dontrun{
cv_mode(dataset = soccer, col_name = result_19_20)}
}
