% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{full_fr}
\alias{full_fr}
\title{full_fr}
\usage{
full_fr(dataset, col_name)
}
\arguments{
\item{dataset}{the dataset for which you wish to retrieve full frequency.}

\item{col_name}{the column name for which you wish to retrieve full frequency.}
}
\value{
returns the full frequency (absolute and relative) for a specified column.
}
\description{
full_fr
}
\examples{
\dontrun{
full_fr(dataset = df, col_name = result_19_20)}
}
