% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{plot_bar}
\alias{plot_bar}
\title{plot_bar}
\usage{
plot_bar(dataset, col_name_1, col_name_2)
}
\arguments{
\item{dataset}{the dataset for which you wish to retrieve the bar plot.}

\item{col_name_1}{the first column name for which you wish to retrieve the
bar plot.}

\item{col_name_2}{the second column name for which you wish to retrieve the
bar plot.}
}
\value{
returns the bar plot for the a variable in different moments.
}
\description{
plot_bar
}
\examples{
\dontrun{
plot_bar(dataset = soccer,
       col_name_1 = result_19_20,
       col_name_2 = result_20_21)}
}
