% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{re_fr}
\alias{re_fr}
\title{re_fr}
\usage{
re_fr(dataset, col_name)
}
\arguments{
\item{dataset}{the dataset for which you wish to retrieve relative frequency.}

\item{col_name}{the column name for which you wish to retrieve relative frequency.}
}
\value{
returns the relative frequency for a specified column.
}
\description{
re_fr
}
\examples{
\dontrun{
re_fr(dataset = soccer, col_name =  result_19_20)}
}
