% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{soccer}
\alias{soccer}
\title{Soccer}
\format{
This data frame contains the following columns:
\describe{
  \item{\code{round}}{The round number.}
  \item{\code{result_19_20}}{The results of the 2019-20 season W - win, L - lose, D - draw}
  \item{\code{result_20_21}}{The results of the 2020-21 season W - win, L - lose, D - draw}
}
}
\source{
\url{()}
}
\usage{
soccer
}
\description{
The results of Sevilla, a Spanish soccer team, during the national league for the seasons 2019-20 and 2020-21.
}
\keyword{datasets}
