% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{tv_d}
\alias{tv_d}
\title{tv_d}
\usage{
tv_d(dataset, col_name_1, col_name_2)
}
\arguments{
\item{dataset}{the dataset for which you wish to retrieve the total variation
distance.}

\item{col_name_1}{the first column name for which you wish to retrieve the
total variation distance.}

\item{col_name_2}{the second column name for which you wish to retrieve the
total variation distance.}
}
\value{
returns the total variation distance for a categorical variable.
}
\description{
tv_d
}
\examples{
\dontrun{
tv_d(dataset = soccer,
   col_name_1 = result_19_20,
   col_name_2 = result_20_21)}
}
